remove Style.hpp
