--- include/Layer/Core/Style.hpp
@@ -1,44 +1,0 @@
-/*
- * Copyright 2023 VeryGoodGraphics LTD <bd@verygoodgraphics.com>
- *
- * Licensed under the VGG License, Version 1.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.verygoodgraphics.com/licenses/LICENSE-1.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-#pragma once
-#include "Attrs.hpp"
-namespace VGG::layer
-{
-
-struct InnerShadowStyle
-{
-  ContextSetting contextSettings;
-  float          blur{ 0.f };
-  Color          color;
-  float          offsetX{ 0.f };
-  float          offsetY{ 0.f };
-  float          spread{ 0.f };
-  bool           isEnabled{ false };
-};
-
-struct OuterShadowStyle
-{
-  ContextSetting contextSettings;
-  float          blur{ 0.f };
-  Color          color;
-  float          offsetX{ 0.f };
-  float          offsetY{ 0.f };
-  float          spread{ 0.f };
-  bool           isEnabled{ false };
-};
-using ShadowStyle = std::variant<InnerShadowStyle, OuterShadowStyle>;
-
-} // namespace VGG::layer
